Extract `MetadataBlock` into dedicated file
