+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/metadata_update_helper.rb
@@ -1,22 +1,7 @@
+require_relative 'metadata_block'
+
 module Fastlane
   module Helper
-    # Basic line handler
-    class MetadataBlock
-      attr_reader :block_key
-
-      def initialize(block_key)
-        @block_key = block_key
-      end
-
-      def handle_line(fw, line)
-        fw.puts(line) # Standard line handling: just copy
-      end
-
-      def is_handler_for(key)
-        true
-      end
-    end
-
     class UnknownMetadataBlock < MetadataBlock
       attr_reader :content_file_path
